Fix serializing of projects object
Previously, an error was generated when trying to return an array of projects from getStaticProps because the members of the array were pending Promise objects. In order to fix that, we can wrap the mapping over those members when creating the array with Promise.all() which will cause each member of the array to resolve before returning the array.