Add hero image to index page header
To make the pages more interesting, use a hero image. There's a separate image for the dark and light modes. The light mode is slightly more complex, because it has to have a filter on it to reduce its brightness so that the white text on top is legible.

The next steps are to also use this on the Projects page and to refactor into reusable components.