Change hero heading to load two next images and hide the opposite color mode
Instead of changing the source of a next/image depending on color mode, which seems to present issues with the blur placeholder, make two next/image blocks, one for each color mode and change their CSS visibility based on the color mode. An initial version using CSS display was unable to hide an image.