Load photos from public directory and add to photos page
Using getStaticProps, get all the photos from the public/images/featured_photos directory and map over them to create an object which is used to add captions, src, height and width to the photos page. The exifr package is used to get the caption. The image-size package is used to get the height and width, as these are absent from EXIF data in photos exported from Lightroom, for some reason.